requirements.txt
fastapi~=0.103.1
python-dotenv~=1.0.0
psycopg2-binary~=2.9.9
uvicorn~=0.23.2